+++ forkbuntu/services/__init__.py
@@ -3,6 +3,7 @@ from .clean import Clean
 from .configure import Configure
 from .extras import Extras
 from .gpg import GPG
+from .initrd import Initrd
 from .pack import Pack
 from .setup import Setup
 from .unpack import Unpack