+++ forkbuntu/steps/__init__.py
@@ -5,9 +5,12 @@ from .create_extras import CreateExtras
 from .initialize import Initialize
 from .load_config import LoadConfig
 from .merge_filesystem import MergeFilesystem
+from .merge_initrd import MergeInitrd
 from .merge_iso import MergeIso
 from .pack_filesystem import PackFilesystem
+from .pack_initrd import PackInitrd
 from .pack_iso import PackIso
 from .sign_iso import SignIso
 from .unpack_filesystem import UnpackFilesystem
+from .unpack_initrd import UnpackInitrd
 from .unpack_iso import UnpackIso