+++ forkbuntu/steps/merge_iso.py
@@ -1,5 +1,4 @@
 from ..step import Step
-from time import sleep
 from munch import munchify
 from os import path
 
@@ -8,9 +7,7 @@ class MergeIso(Step):
         'past': 'merged iso',
         'present': 'merging iso'
     })
-    requires = [
-        'load_config'
-    ]
+    requires = ['load_config']
 
     def run(self):
         s = self.app.services