+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.3] - 2018-06-01
+### Added
+* modify initrd filesystem
+### Fixed
+* empty checksum bug
+
 ## [0.2.2] - 2018-06-01
 ### Added
 * caching
@@ -13,7 +19,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 * chown bugs
 * postinstall script
-
 
 ## [0.2.1] - 2018-05-31
 ### Added